Moved some utility functions to new Tools class, will later move time specific functions to new class
